requirements.txt
Django==4.1.4
envparse==0.2.0
psycopg2-binary==2.9.5
asgiref==3.5.2
pip==21.3.1
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1
djangorestframework==3.14.0
social-auth-app-django==5.0.0
django-cors-headers==3.13.0
